revise json file loading for tests
With asdf-binary-locations on, as it will be in new versions of ASDF,
it is no longer possible to use load-pathname or load-truename to find
files from loaded binaries.  In the general case, binaries may not be
located in the same directory as the source files from which they are
compiled.

We avoid this problem by caching the value of *load-truename* inside
code that is run in the .asd file, and then referring to it later on.

The *load-truename* of the .asd file will remain constant.

darcs-hash:20100102214917-b37ea-e460cd8b226ae1faa5ef7a2dd62d04854467737e.gz